Added left/right masking Moved statics to separate file
